Added ConsoleTesting project to solution
ConsoleTesting contains several test functions for testing methods to be
used in D3D9Callback as well as analyze tonberry textures.

Old_Hash_Algorithm_1: Omzy's original algorithm modified to use cv::Mat

Hash_Algorithm_1: Omzy's algorithm optimized with a static array of
coordinates used in the hash

FNV_Hash: a fast hashing algorithm that takes octet pixel values and spits
out a 64-bit hash. It is comparable to Omzy's in speed but uses actual
values instead of pairwise comparisons. FNV_Hash takes an array of
coordinates as input, and if desired adds the average R, G, and B values
to the end of the hash to distinguish identical but different-colored
images

Copy_Unique_Left_Half: use md5 hashes to find the unique left halves of
the textures in all debug\ folders that don't end in '0', then save those
unique left halves as new textures in dest

Copy_Unique_Left_Objects: use md5 hashes on the output of
Copy_Unique_Left_Half() to find unique top-half and bottom-half textures
and save them to dest

Analyze_Pixels:

1) first divides the image into 10x10 blocks with a 1-pixel
border in between blocks, then finds the pixel coordinate in each block
that has the highest local variance. The 1-pixel padding is to prevent
neighboring coordinates from being selected. The resulting coordinates are
used with the FNV Hash and hash collisions are recorded.

2) uses the collisions from part 1 to find the 44 most
frequently-differing pixel coordinates amongst the colliding hashes, then
adds these to the coordinates used for hashing. The result is again used
with the FNV Hash and collisions are recorded.