D3D9Callback using new cache system
While JeMaCheHi created a HashCache and a TexCache in cachemap.h/cpp, only
the HashCache is being implemented in this version. The TexCache is
implemented via global variables in GlobalContext.cpp, based on the
original collabedit implementation.

The newhandle cache holds replacement textures and keeps track of their
access order. It uses a Least-Recently-Accessed list combined with a
hashmap to map hash :-> newhandle. The hashmap (nh_map) maps uint64 hashes
to iterators (pointers) to entries in the LRU list (nh_list). Entries in
nh_list are, in turn, a pairing of the hash and the HANDLE pointer, and
the list is singly-linked to allow O(1) insertion at the beginning,
deletion from the back, and transfer of any item to the front of the list
whenever it is accessed.

Newhandle Cache Acess:
 - find the hash in nh_map, which gives pointer to nh_list item
 - move nh_list item to front of list
 - point nh_map[hash] at nh_list.begin()

Newhandle Cache Insertion:
 - push (hash, HANDLE) pair to the front of nh_list
 - if nh_list is greater than the allowed cache size, get the last item
 - release the last item's HANDLE, erase the hash from nh_map, and pop it
     from the back of nh_list

Paired with the nhcache is another unordered map that maps the in-game
texture HANDLES to iterators (pointers) to entries in the nh_map. When a
HANDLE's hash is not found, the HANDLE is erased from the handlecache in
case that memory location had previously held a replacable texture. If
there is a hash match, the newhandle is inserted (or looked up) in the
nhcache and handlecache maps the HANDLE to the nhcache entry. In
SetTexture(), HANDLES are looked up from the handlecache to the nhcache
and, if a match is found, the in-game texture is replaced appropriately.