New hashing process now working in-game
The hash algorithm itself was modified to use the pixels with the smallest
mode count (i.e. the number of images that have the most common pixel
value [0-255]) instead of those with the maximum variance. This is a more
accurate value of variance when using discrete values. The result was
fewer collisions, even without doing the collision analysis. Collision
analysis is still a potential improvement but the process needs to be
improved so that all the frequently-colliding pixels are not right next to
each other.

FNV_Hash_Combined now generates up to three hashes for a texture. If the
texture has height <=128, the result is a single FNV hash. However, for
textures with height >128, we generate a hash_combined, a hash_upper, and
a hash_lower:

 - hash_combined: self-explanatory, just one long hash using the analysis
   coordinates on the upper left object and then the lowest left object,
   stopping at the Rect: (0, 128, 128, 256). Everything to the right of
   x=128 and below y=256 is ignored.

 - hash_upper: what the combined_hash would be if every pixel below the
   upper left object was (0,0,0). This is calculated quickly by
   multiplying the hash of the upper left object by
   FNV_NOLOWER_{RGB_}_FACTOR = (FNV_OFFSET_PRIME)^FNV_COORDS_LEN

 - hash_lower: like hash_upper, this is what the combined_hash would be if
   every pixel above the lower left objcet was (0,0,0). This is calculated
   along with the second half of hash_combined by starting with
   FNV_NOUPPER_{RGB_}_BASIS = FNV_OFFSET_BASIS * FNV_NOLOWER_{RGB_}_FACTOR

A complicated fieldmap data structure was created such that fieldnames
only ever exist in memory once. Basically instead of having:

    std::unordered_map<uint64, std::unordered_set<string>> fieldmap

We have:

    std::unordered_map<uint64, std::unordered_set<string*>> fieldmap

While the amount of memory saved is relatively small, it's a clever
solution and it made me happy so I created it.

In UnlockRect(), the logic is thus:

 1) If nhcache[hash_combined] exists, USE IT; else:
 2) If fieldmap[hash_combined] exists, CREATE IT AND USE IT; else:
 3) If nhcache[hash_upper] exists, USE IT; else:
 4) If fieldmap[hash_upper] exists, CREATE IT AND USE IT; else:
 5) If nhcache[hash_lower] exists, USE IT; else:
 6) If fieldmap[hash_lower] exists, CREATE IT AND USE IT; else:

Note that there is still an issue with multiple objects on the same
in-game texture when the objects come from different replacements. In the
rare case that two different objects that are rendered with the same
palette appear together, only the upper object will be replaced. If the
rendering palettes differ, only 1 object will correctly hash to a
replacement, so no worries. In future versions, we will need to be able to
load and combine object replacements from 2 different .PNG images.

Finally, an issue with the nhcache needed rectifying; whenever entries in
the nhcache were modified, either by moving items in the nh_list or
erasing items from the nh_map, there was the possibility that entries in
the handlecache that pointed at the nhcache entries would become
invalidated. The reverse_handlecache (nh_map_iter :-> Handle) was added so
that handlecache entries could be found and removed in these situations.
handlecache is now essentially a doubly-indexed map.