Fixed cache problems; now officially working
We were dumb in the way we constructed the hash: unordered_sets and
unordered_maps can invalidated their iterators when inserts or erasures
cause a rehash, which keeps the number of hash buckets low. Because of
this, mapping replaced HANDLEs to nh_map::iterators could lead to problems
when nh_map was rehashed. But moreso than that, it was needless to store
the nh_map::iterators in the first place. We should've simply been mapping
to the uint64 hashes, since the nh_map entry is only 1 O(1) map lookup
away.

Additionally, the reverse_handlecache needed to be a multimap; a replaced
HANDLE only maps to one replacement newhandle, but many different replaced
HANDLEs can map to the SAME replacement newhandle.

So handlecache_t was changed to std::unordered_map<HANDLE, uint64> and
reverse_handlecache_t to std::unordered_multimap<uint64, HANDLE>. Using
hashes instead of iterators removed the need to update other handlecache
entries when the nh_list entries were re-ordered (and subsequently, nh_map
entries were changed) because there was nothing to invalidate--the hash
stays the same, and the hash lookup still returns the right nh_list item.

Also, erasures from the nh_list now correctly cause all nh_map items that
pointed to the erased replacement newhandle to be removed by iterating all
hash matches in the reverse_multimap.

Of course, the realizations about iterator invalidated in unordered
containers all was cause for concern in the FieldMap structure. This was
solved by changing the fieldset to an ordered set, such that iterators
won't be invalidated. While fieldset access will be slightly slower, there
are relatively few fields so this should not be a problem.
I also discovered that std::hash<std::string> is a
pretty horrible function (with lots of collisions), so the hash of the
fields' pointers is now used in the fieldset_iter_set instead of the hash
of the strings themselves.
NOTE: there may still be problems with this setup, but I'm fairly sure it
is very unlikely to fail, could only possibly fail in the event that there
are WAY more fields than we are currently working with, and is not
something we need to worry about

TODO: implement SetTexture() for merging existing newhandles