Fixed texture cache erasure errors
I was iterating over reverse_handlecache.equal_range(hash) and erasing
iterators within the loop, which caused problems. Now I get the HANDLEs to
be deleted from handlecache using the equal_range, then call
reverse_handlecache.erase(hash) after the loop.

Also, there is a rare (and, honestly, inexplicable) case wherein
TextureCache::at(HANDLE) finds the HANDLE in the handlecache but the hash
it's mapped to does not exist in the nh_map. This should never happen
normally, but catching it and printing the handlecache entry in question
reveals that this only ever happens at the very beginning of the log file.
There's something fishy going on here, but dealing with that edge case
causes the handlecache to perform beautifully.

Debug lines were also added to create_newhandle(), but they proved to be
unnecessary. As far as I can tell, create_newhandle() is working
perfectly.

TODO: still need to add newhandle creation from upper and lower halves of
existing newhandles