cache error fixed, GlobalContext given FNV_Full()
GlobalContext now uses a FNV hash of every pixel in the image to determine
whether an in-game texture has already been saved to the \debug\nomatch
folder. This results in very few collisions, but possibly more saved files
than we really want.

When insertions are made to the cache map that change the replacement
mapped to a replaced HANDLE that already existed in the cache (i.e. the
replaced HANDLE was re-used by the game), the stale reverse_handlecache
entry needed to be removed before the new one was added.

In fixing this cache error, it was discovered that reverse_handlecache
entries seem to be randomly disappearing. This is not good. The only
explanation I can come up with is that the uin64 hashes are causing
collisions internally when the map is rehashed. I'm going to try to switch
to a uint32 to see if that helps.