All known TextureCache bugs fixed. Works perfectly
I was dumb and forgot I was using the ::erase() function (to remove
handlecache entries for HANDLEs that have been recycled by the game), and
didn't realize it was erasing all reverse_handlecache entries for a hash
instead of just the one pointing towards the HANDLE to be erased. Fixing
this issue resolved the entries-disappearing error I described in the last
commit message.

NOTE: it might be worth it to detect when the reverse_handlecache entry
that is erased is the LAST entry for a particular hash, then to move that
hash's nh_list entry to the back of the list. This way nhcache entries
that are not currently pointed at by any HANDLEs in the handlecache will
be removed first in the event that the cache grows too large. Something
to think about... do the performance gains weight out???

Also, I added a FNV_Hash_32 and associated constants to ConsoleTesting
when I was trying to determine if the cache bug I just fixed was due to
hash collisions in unordered_multimap<uint64, HANDLE>. It's no longer
being used, but it's there if we ever need it.