Project organization; more pixel analysis
Removed some nonessential libraries from D3D9CallbackSC2/lib (now in
/lib_bak).

Added boolean options to Analyze_Pixels to toggle various tests. Also
added struct image with mat() function to imread image off of HDD and
return a rect representing a <=128x128 object within the image. This is
because there are too many images (~17,000) in analysis0 to hold in
memory, so they have to be read over and over.

TODO:
 - run Algo_1, Highvar, and Lowmode all at once to avoid reading images in
   over and over
 - use every-other-row/col and ranked results instead of highest-in-block
   for highvar/lowmode