Use lowmode from collision in 1000 rounds
Adds coords to those used in the FNV hash one by one, looking through
COLLISIONS ONLY for the lowest mode and adding that to the coords deque
and then, using the new coords, hash, search for collisions, rinse repeat.
The buffer was too small to see all 1000 rounds but when using only the
~6500 images that exclude backgrounds, collisions narrowed in on ~900
collisions before 700 iterations.

This, as it turns out, is not the best strategy, though it beats me as to
why.

TODO:
 - modify the coord struct to use a color enum as well as x and y
 - hash all colors for all coords individually AND/OR
 - use individual (x,y,c) coords in the hash, instead of all 3 colors for
   each pixel