Use individual RGB in progressive lowmode search
Starting with every image in the collision set, search for the coordinate
with the lowest mode (individual x,y,RGB) and add it to the set of
coordinates to use in the hash. Then run the hash to get a new collision
set and iterate.

After 500 iterations the latest verison results in about 800/6923
collisions. Analysis of these collisions shows that many colliding pairs
have 0 differing pixels, which means the FNV Hash on the full set of
coordinates is inadequate for eliminating duplicates in the initial
imageset.

TODO: implement Murmur2 hash instead of FNV