Fix collisions/hash2map problem
The Matchtype variable match, which holds the return values from
getfield(), was not getting the return value from getfield2(). This was a
simple typo that broke the collisions resolution logic.

Additionally, the hash type of Hash_Algorithm_2 is a BigInteger and thus
was not being stored in the cache--rather, the original collision hash was
stored. A simple StringToUint64() method was added to convert these hashes
into uint64's equal to BigIntegerHash % UINT64_MAX.

TODO:
 - check to make sure the converted hashes don't collide with existing
hashes
 - figure out how JeMa makes d3d9Callback.dll so damn small