Fixed broken collisions / hash2map
We were not properly saving the matchtype enum from getfield2, so there
would never be hash2 matches and thus all colliding textures were broken.
What's more, the hash2 hashes are BigIntegers and cannot go in the texture
cache as such, so textures would be replaced incorrectly when they matched
a colliding hash.

I added a StringToUint64 that converts BigIntegers to uint64_t, made sure
the converted hash2 values didn't collide with any known mod hashes, then
used those values to store collision replacements in the texture cache.

TODO: fix the unknown exception crash; for some, disabling Project Eden
resolves the issue so at first I thought it was still a collision issue,
but now some people are saying it crashes regardless. Because it is a
"random crash" I suspect it happens when the cache is full. This is a very
high priority issue.