Enable debug logs; fix sysfld/iconfl functions
If smaller textures were colliding with sysfld/iconfl hashes, there's a
possibility that the special-case functions would look for a pixel outside
of the size of the HANDLE. I'm not sure that this is what was causing the
crash, but it's possible. I also switched to parsing the texture number
for creating the hash in parse{sysfld,iconfl}() and using a switch
statement instead of nested if/else string comparisons. The
global_context.log made it apparent that these functions are being called
far more frequently than I previously thought so those string comparisons
were really hurting us.

Decreasing the size of the cache to minimum (100) did not cause any
crashes for me, so I'm not so sure it's a cache problem anymore. Hopefully
some qhimm players will use the 2.04_ALPHA version and the logs they
collect will shine some light on the issue.