Crash resolved - it was an out-of-memory exception
I messed up my calculations when determining optimal cache size. Each
IDirect3DTexture9 pixel is 4 bytes, so replacement textures are typically
1024x1024x4=4MB and thus only 512 would fill 2GB of RAM. We have to leave
space for FF8_EN.exe, so a safe default cache_size is 250. If the user
enables LargeAddressAware on FF8_EN.exe, we can use 450-500 textures with
3GB and 700-800 with 4GB. The largest cache_size I have tested is 800, and
even then it may have crashed after a longer play time.

This commit includes lots of extra debugging in the texture creation
process that should be removed to prevent speed losses (albeit minor
ones).

In TextureCache, nhcache_list items are now spliced to the back of the
list instead when a HANDLE pointing to them is erased. This is in an
effort to prevent premature deletion of still-referenced cache items by
instead deleting unreferenced cache items. Note that nhcache_list items
that are pointed at by two different handlecache entries will still be
moved to the back when only one of those references is erased. Checking
for other references should be possible using the reverse_handlecache.

TODO:
 - move settexture debug conditionals inside preprocessor macro for
   optimal speed
 - don't splice still-referenced cache items to the back of the list, or
   remove this functionality completely
 - create a persistent list from which entries cannot be removed, and
   allow modders to specify essential textures (like iconfl/sysfld)
 - possibly preload the essential textures?