Use item bool to hold persist instead of set
Instead of a std::pair<uint64_t, HANDLE>, nhcache_item_t is now a struct
with an additional boolean value designating whether the item should
persist or not. This prevents the overhead of a unordered_set lookup when
determining whether to move the item after an at() call (fastupdate()), as
well as in several other places.

Additional logic was also added to allow for arbitrary cache sizes,
although the lag is exponentially worse with cache sizes below 100. Now if
there are too many persistent textures, Tonberry won't crash... it will
just lag.