This is the release version of 2.04
I'm returning to this after 3 months of inactivity, so I'm not sure
exactly what the differences are with the last commit. I'm fairly
confident that this is exactly what was released as v2.04, though most of
the changes appear to be mostly aesthetic (indention levels, etc.)

Note that some notable changes to the cache need to be migrated to the
master branch.